Screenlog management on server deletion
As per what @rocco27 suggested awhile ago, this adds some sort of
control over servers that are about to be deleted ( See:
http://www.opengamepanel.org/forum/viewthread.php?thread_id=3214#post_19798
)

It adds 3 button elements. Clicking on them dynamically updates the href
value of the 'Yes and delete the files' link' - adding another GET
variable for parsing.

It also includes what @Zorrototo suggested for the ability to backup the
log files, perhaps for later inspection.

Tested on Ubuntu and Windows, works on both.