Minor Fixes
This is mostly fixes for warnings or notices that appear with max error
reporting. There's still othr stuff, but this'll do for now.

Form_table_class: Uses the old, PHP-4 constructor method which is now
deprecated.
Refreshed: Same as above, used the PHP-4 constructor method. PHP 5.3 and
above is supported. So, no need to use much, much, older method.

lib_remote: Set the default timeout in the constructor method to match
that of what was commited in 6e1629f , and changed the remote_dirlistfm
method from having a function within the public method to utilize the
callback arguement.

view: Added a better check for the installation file removal. $admins
will either be false or an empty array. This prevents the script
attempting to delete the file every page load if the file doesn't exist.

show_users: Added single quotes around the user_id key to prevent a
notice.

TS3Admin: getRemoteServers returns false is there's no servers. Added a
check to make sure there is remote servers - if there's not, display a
message. Added a translation.

FTP Users: getRemoteServers returns false is there's no servers. And as
such you cannot add FTP users. So display a message to the user
notifying them about this.

is_numeric being passed to array_walk in this fashion was always
returning true, however generates muliple warnings. Now, array_filter is
used and the count value of that is used. Which also removes the need
for the first check.

Status: Added a check for the GD extension for the bar.php images. If GD
isn't loaded, the images won't be displayed but the other components
are. A message will be displayed telling the user that GD is required.
Added translations.