Assigning Multiple Cores to Servers
Fulfilling a feature request posted
[here](http://www.opengamepanel.org/forum/viewthread.php?thread_id=5285)
by @FarukGamer - these small changes allow you to assign multiple cores
to a server.

It requires a change in the ogp_game_mods table, changing the
cpu_affinity column from a varchar(2) to varchar(64) (a maximum possible
of 32 cores, as the core info is seperated by commas, eg, 0,1,2)

It doesn't require anything to be changed on the agent side - as the
agent just uses `taskset`, which accepts cores the way I mentioned
above.

It'll also attempt to verify that a valid core has been selected. As in,
if you've got 4 cores, the user cannot edit the checkboxes and change
the value of one to a non-existing core. Or edit the checkbox value to a
non-numeric value. Doing so will default the field to 'NA', which the
agent handles.

It's mostly srcds servers I've hosted for in the past which don't
benefit from multiple cores. Although, I could see this being useful for
some games.