Remember Checked Affinity Boxes.
As per @FarukGamer and @Zorrototo's suggestion, this pull request will
put the 'checked' attribute on each checkbox next to the CPU checkboxes
which are currently assigned to a server.

I added a new database method so the data may be used elsewhere if so
desired.

Tested on Windows and Linux Mint.