ARK: Survival Evolved xml update
Finally it is here. I rebuilded the whole xml files from the scratch and added most of the important server and game settings.
No changelog this time, because it was too long and GitHub was not able to manage it. :-)

I tested (also use) only Linux version of the config. I also made Windows version, but it doesn't work for me under WINE. So feedbacks are welcome about does it work or not.