Minor Fixes, Pagination Improvements, Check DB Connection.
Home and Index attempt to execute as normal if there's no connection to
the database. There's a language translation in global.php
`no_db_connection` which can be used if there's no connection, we just
need to check for an instance of OGPDatabase.

As noted
[here](http://opengamepanel.org/forum/viewthread.php?thread_id=5329#post_25965)
with the pagination limits applying to the menu, that's been fixed.

A try/catch block was added to update/update.php and extras/extras.php
in order to prevent what happened
[here](http://opengamepanel.org/forum/viewthread.php?thread_id=5328)

A function was added to functions.php due to the excessive code reuse to
render the pagination page numbers. Additionally, the previous
pagination pages were a bit bugged - i.e, if you've got 3 items, set
$limit to 2, you'd see 4 pages.

The new function also allows each page number to be styled in themes, as
it's possible to pass a customClasses array to it (as in, for CSS)