Pagination Improvements
Implements the suggestion by @Obada8 made
[here](http://www.opengamepanel.org/forum/viewthread.php?thread_id=5329#post_25974).

This will also make for cleaner output in the event there's several page
links that need to be rendered since the maximum that can be shown is
defined by `$pagesShown` in the `paginationPages` function.

I'm no longer passing an array to the paginationPages function but
rather just a class prefix.

You can see in the attached image how page numbers are rendered now.
There's a total of 11 pages - the first and last page will always be
visible. The current page number will always be in the middle. To show
more page numbers, change the `$pagesShown` argument passed to
paginationPages.

![Pagination Image](http://i.imgur.com/HeziaMH.png)