Missing a strings from the resource
Actually we already lost 2 strings from the english language before. It happend by an accident. Transifex synced back to GitHub, but the Transifex was not up-to-dated yet this time. Auto sync all resource files on Transifex (from GitHub to Transifex) takes ca. 1-2 days to recheck, compare all files and update which needs to. Checking time can not be change and no option to force it too unfortunatelly.

Link: https://github.com/OpenGamePanel/OGP-Website/commit/e34c585d2a5c2e7336707ac28d51e5bd8e20dc57#diff-239db8f6b0cf4bd1d30c144c643f833f