TS3 Module Fixes
Updates the Smarty template engine, and also set the template path in
ts3webinterface.class.php - as Smarty wasn't able to find the
directories previously.

Also added a port field in the database instead of trying to connect to
the default 10010 port so servers running on non-standard ports
shouldn't cause errors when attempting to use the module.