Search Field and other minor fixes
Fixes the issues noted by Omano in #180.

Also now checks if php-zip is installed on the update module.
Previously, if you attempted to update without this extension you'd
receive a warning.

Additionally fixed not being able to set an expiry date on user accounts
as mktime() was returning false on PHP >=7 due to the $is_dst arg being
removed - and had been deprecated since 5.1.0.