IV Multiplayer (GTA IV) xml update
Before I used IVMP 0.1 T3 version, but had a lot of problems. The server worked, but the client did not, Then I found the latest public version of it, called IVMP 0.1 RC2 and finally tested the server with client together. Everything works fine.
The xml files are done. I changed and increased only the maxplayers in this update, since the changelog said this could be 48 maximum in this version.

Well this project is closed and seems nobody interested for this modification anymore, but anyway I upgraded it, because I wanted to test it how it works.