Reign of Kings xml update
Windows 64 bit support of this game added to OGP.
The xml file manages all possible things execpt log file (server uses unique file names always), the LGLS query does the same as Steam and read wrong infomations (I guess there is a problem with the query protocol in the server, I also wanted to have no query, but then I can not manage the query port in the config file).
The "max_user_amount" is fixed in 64. It can be higher, but the typical player slots are 20-50, if it is too high then the server gets lagg.