FiveM XML
Requested here:
https://www.opengamepanel.org/forum/viewthread.php?thread_id=5747

A key is required to run servers, which can be generated at
https://keymaster.fivem.net/

The IP / Port the server binds to is specified in the server.cfg file.
So cfg_text_replace was changed so the IP / Port may be retrieved via
the ip_port key.

The post install script handles the installation of the server. It'll
pull the newest version of the server for the appropriate platform.