duplicated lang constant, hasValue to isset and !empty, function getThemePath
There is a language constant defined in global.php for no_remote_servers.
PHP warns about hasValue all over the site, replaced all that throw warnings that I found.
Added a function to get the theme path, so every other place where this is needed will get the same path than the global, this avoids an error where, if the user has configured a different theme than the panel settings, it was showing the images for the panel settings theme instead of the images of the theme selected by the user.