Fixed a problem with server startup
Replaced the loop of $_REQUEST inside of the params loop by an isset-POST conditional, because the loop was unnecessary, also check if the parameter is locked by admin before forcing it to be present. Also save the original value of checkboxes instead of true, because the comparsion on the next startup would not detect it as a checkbox value since the key and the value doesn't match. And changed {DEPENDS:other_game_server_path_additional} to {DEPENDS:$paramType}, otherwise it would allways detect it as "other_game_server_path_additional" when parsing the POST key.