Changes (on new setups) the charset of the database from latin1 to utf8mb4
Changes the default charset to utf8mb4 to make possible storing unicode characters.
It will only affect to new setups.

Why is it necessary? the main reason is to avoid the following error when searching for in the database:
**Illegal mix of collations (latin1_swedish_ci,IMPLICIT) and (utf8mb4_general_ci,COERCIBLE) for operation 'like'**

The update module contains the necessary queries to update a database (in modules/update/module.php), but it might create other problems after the conversion, so that is commented. By uncommenting it, changing the database version to 3 and using the "update modules" button in the panel, it will (try to) convert the charset of the database.